k8s_cp turbo mode compliance
